miscellaneous bits: change flags to an unsigned long and totsize to a size_t, store the number of multibyte characters instead of the number of single-byte characters in totsize, and add a few formatting fixes
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2302 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/nano.h
@@ -172,8 +172,8 @@ typedef struct openfilestruct {
     size_t placewewant;		/* Current file's place we want. */
     int totlines;		/* Current file's total number of
 				 * lines. */
-    long totsize;		/* Current file's total size. */
-    long flags;			/* Current file's flags: modification
+    size_t totsize;		/* Current file's total size. */
+    unsigned long flags;	/* Current file's flags: modification
 				 * status (and marking status, if
 				 * available). */
     file_format fmt;		/* Current file's format. */
